+++ packages/ui/vite.config.js
@@ -3,16 +3,25 @@ import { resolve } from 'path'
 import dts from 'vite-plugin-dts'
 
 export default defineConfig({
-    plugins: [dts({
-        insertTypesEntry: true,
-        include: ["components/**/*", "index.ts"],
-        rollupTypes: true,
-        outDir: 'dist',
-    }
-    )],
+    plugins: [
+        dts({
+            insertTypesEntry: true,
+            include: ["components/**/*", "index.ts", "css.d.ts"],
+            rollupTypes: true,
+            outDir: 'dist',
+        }),
+    ],
+    css: {
+        modules: {
+            generateScopedName: 'retro-ui__[name]__[local]__[hash:base64:5]',
+            scopeBehaviour: 'local',
+            localsConvention: 'camelCase',
+        },
+    },
     build: {
         lib: {
             entry: resolve(__dirname, 'index.ts'),
+            include: ['components/**/*'],
             formats: ['es', 'cjs'],
             name: 'retro-ui', // Global variable name when used in browser
             fileName: (format) => {
@@ -21,6 +30,8 @@ export default defineConfig({
                         return 'index.mjs'
                     case 'cjs':
                         return 'index.cjs'
+                    default:
+                        return 'index.js';
                 }
             }
         },
@@ -34,5 +45,7 @@ export default defineConfig({
                 }
             }
         }
-    }
+    },
+    assetsInclude: ['**/*.ttf'],
+
 })