+++ apps/web/CHANGELOG.md
@@ -1,5 +1,32 @@
 # web
 
+## 0.3.0
+
+### Minor Changes
+
+- docs added with styleguidist + refactor of components
+
+### Patch Changes
+
+- Updated dependencies
+  - @duck4i/retro-ui@0.3.0
+
+## 0.2.3
+
+### Patch Changes
+
+- basic components and styles implemented
+- Updated dependencies
+  - @duck4i/retro-ui@0.2.3
+
+## 0.2.2
+
+### Patch Changes
+
+- first components and UI
+- Updated dependencies
+  - @duck4i/retro-ui@0.2.2
+
 ## 0.2.1
 
 ### Patch Changes