+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @duck4i/retro-ui
 
+## 0.3.0
+
+### Minor Changes
+
+- docs added with styleguidist + refactor of components
+
+## 0.2.3
+
+### Patch Changes
+
+- basic components and styles implemente
+
+## 0.2.2
+
+### Patch Changes
+
+- first components and UI
+
 ## 0.2.1
 
 ### Patch Changes